TF backend: remove 'common' submodule
